refactor: resolve cyclic dependency in cui-jwt-validation package
Move ValidationContext from pipeline to domain.context package to eliminate
circular dependency between pipeline and domain.token packages.

The issue was:
- ExpirationValidator (pipeline) → TokenContent (domain.token)
- TokenContent (domain.token) → ValidationContext (was in pipeline)
Creating a cycle: pipeline ↔ domain.token

Solution:
- Moved ValidationContext to de.cuioss.jwt.validation.domain.context
- Updated all imports in main and test files
- Added package documentation for new domain.context package

This refactoring improves the architecture by placing ValidationContext
in a more appropriate location that both pipeline and domain.token
packages can depend on without creating cycles.
