+++ cui-jwt-validation/src/main/java/de/cuioss/jwt/validation/TokenValidator.java
@@ -18,6 +18,7 @@ package de.cuioss.jwt.validation;
 import de.cuioss.jwt.validation.cache.AccessTokenCache;
 import de.cuioss.jwt.validation.cache.AccessTokenCacheConfig;
 import de.cuioss.jwt.validation.domain.claim.ClaimValue;
+import de.cuioss.jwt.validation.domain.context.ValidationContext;
 import de.cuioss.jwt.validation.domain.token.AccessTokenContent;
 import de.cuioss.jwt.validation.domain.token.IdTokenContent;
 import de.cuioss.jwt.validation.domain.token.RefreshTokenContent;
@@ -34,7 +35,6 @@ import de.cuioss.jwt.validation.pipeline.TokenBuilder;
 import de.cuioss.jwt.validation.pipeline.TokenClaimValidator;
 import de.cuioss.jwt.validation.pipeline.TokenHeaderValidator;
 import de.cuioss.jwt.validation.pipeline.TokenSignatureValidator;
-import de.cuioss.jwt.validation.pipeline.ValidationContext;
 import de.cuioss.jwt.validation.security.SecurityEventCounter;
 import de.cuioss.tools.logging.CuiLogger;
 import de.cuioss.tools.string.MoreStrings;
