+++ cui-jwt-validation/src/main/java/de/cuioss/jwt/validation/domain/context/ValidationContext.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package de.cuioss.jwt.validation.pipeline;
+package de.cuioss.jwt.validation.domain.context;
 
 import lombok.Getter;
 import lombok.NonNull;
